connect database and restructure the code
