+++ .gitignore
@@ -1,5 +1,12 @@
 app/__pycache__/
 app/api/__pycache__/
 app/api/routes/__pycache__/
+app/utils/__pycache__/
+app/services/__pycache__/
+app/schemas/__pycache__/
+
+
+
+
 .env
