+++ app/api/routes/__init__.py
@@ -1,7 +1,7 @@
 from app.api.routes.login import *
 from app.api.routes.order import *
 from app.api.routes.status import *
-from app.api.routes.user import *
+# from app.api.routes.user import *
