+++ app/main.py
@@ -1,6 +1,13 @@
 from fastapi import FastAPI
-from .api import api_router
+from app.api import api_router
+from contextlib import asynccontextmanager
+from app.database import init_db
 
-app = FastAPI()
+@asynccontextmanager
+async def lifespan(app: FastAPI):
+    init_db()
+    yield
+
+app = FastAPI(lifespan=lifespan)
 
 app.include_router(api_router, prefix="/api/v1")
