+++ docker-compose.yml
@@ -1,0 +1,35 @@
+services:
+  api:
+    build: 
+      dockerfile: Dockerfile  
+    env_file:
+      - .env     
+    ports:
+      - "8000:8000"  
+    volumes:
+      - ./app:/app
+
+  db:
+    image: postgres:16-alpine
+    environment:
+      POSTGRES_USER: "postgres"
+      POSTGRES_PASSWORD: "admin" 
+      POSTGRES_DB: "mydb" 
+    ports:
+      - "5432:5432"  
+    volumes:
+      - pgdata:/var/lib/postgresql/data
+  
+
+  pgadmin:
+    image: elestio/pgadmin:latest  
+    environment:
+      PGADMIN_DEFAULT_EMAIL: "admin@admin.com"  
+      PGADMIN_DEFAULT_PASSWORD: "admin"  
+    ports:
+      - "5050:80"  
+    depends_on:
+      - db
+
+volumes:
+  pgdata:
